Split operator route test suites
