+++ apps/operator-ui/src/app/OperatorRoutes.controlPlane.mutations.testSuite.tsx
@@ -1,0 +1,11 @@
+import { registerOperatorRoutesControlPlaneActionRequestTests } from "./OperatorRoutes.controlPlane.actionRequests.testSuite";
+import { registerOperatorRoutesControlPlaneCaseRecordTests } from "./OperatorRoutes.controlPlane.caseRecords.testSuite";
+import { registerOperatorRoutesControlPlanePromotionTests } from "./OperatorRoutes.controlPlane.promotion.testSuite";
+import { registerOperatorRoutesControlPlaneTicketRequestTests } from "./OperatorRoutes.controlPlane.ticketRequests.testSuite";
+
+export function registerOperatorRoutesControlPlaneMutationTests() {
+  registerOperatorRoutesControlPlanePromotionTests();
+  registerOperatorRoutesControlPlaneCaseRecordTests();
+  registerOperatorRoutesControlPlaneActionRequestTests();
+  registerOperatorRoutesControlPlaneTicketRequestTests();
+}
